Recursive torrent download using gateway
